New user-article relation

+++ app/controllers/tweets_controller.rb
@@ -10,11 +10,14 @@ class TweetsController < ApplicationController
   end
 
   def new
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
+
     @tweet = Tweet.new
   end
 
   def create
     @tweet = Tweet.new(tweet_params)
+    @tweet.user = current_user
 
     if @tweet.save
       redirect_to @tweet
@@ -25,7 +28,9 @@ class TweetsController < ApplicationController
   end
 
   def edit
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
     @tweet = Tweet.find(params[:id])
+    session_notice(:danger, 'Wrong user!') unless equal_with_current_user?(@tweet.user)
   end
 
   def update
@@ -38,10 +43,16 @@ class TweetsController < ApplicationController
   end
 
   def destroy
-    tweet = Tweet.find(params[:id])
-    tweet.destroy
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
 
-    redirect_to tweets_path
+    tweet = Tweet.find(params[:id])
+    
+    if logged_in?
+      tweet.destroy
+      redirect_to tweets_path
+    else
+      session_notice(:danger, 'Wrong user!')
+    end
   end
 
   private
